% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGRangesObjectPMDSegNew.R
\name{createGRangesObjectPMDSegNew}
\alias{createGRangesObjectPMDSegNew}
\title{Create an GRanges Object containing the PMD segmentation.}
\usage{
createGRangesObjectPMDSegNew(m, y.list, num.cores=1, seqLengths)
}
\arguments{
\item{m}{A GRanges object containing the coordinates, total (T) and methylated counts (M)}

\item{y.list}{A list of the predicted results.}

\item{num.cores}{The number of cores used for the calculations (default 1)}

\item{seqLengths}{A named vector indicating the chromosome lengths of the genome used.}
}
\value{
A GRanges object containing the PMD segmentation.
}
\description{
Create an GRanges Object containing the PMD segmentation
}
\examples{
\dontrun{
library("MMSeekR.data")
library("MMSeekR")
data("NNscore.hg19")
data("hg19.seqLengths")
data("hg19.blackList")
methFile <- system.file("extdata", "TCGA_BLCA_A13J_test.tab", package="MMSeekR")
meth <- readMethylomeNew(fileName=methFile, NNdat=NNscore.hg19, seqLengths=hg19.seqLengths)
if(!dir.exists("TestResult/")){dir.create("TestResult/")}
hmm.modelList=trainPMDHMDNew(meth, "chr16", 201, 1, "TestResult/TCGA_BLCA_A13J_test.multiModel.pdf")
y.list=PMDviterbiSegNew(meth, hmm.modelList, 201, 1)
segments = createGRangesObjectPMDSegNew(meth, y.list, 1, hg19.seqLengths)
}

}
