% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMDviterbiSegNew.R
\name{PMDviterbiSegNew}
\alias{PMDviterbiSegNew}
\title{Predict PMD and HMD segments}
\usage{
PMDviterbiSegNew(m, hmm.modelList, nCGbin=201, num.cores=1)
}
\arguments{
\item{m}{A GRanges object containing the coordinates, total (T) and methylated counts (M)}

\item{hmm.modelList}{A list of the parameters of the HMM model}

\item{nCGbin}{The number of CpGs in each sliding window used to calculate PCC correlation and Mvalue (default 201). The default is highly recommended.}

\item{num.cores}{The number of cores used for the calculations (default 1)}
}
\value{
A list of the predicted results.
}
\description{
Predict PMD and HMD segments using the Viterbi algorithm
}
\examples{
\dontrun{
library("MMSeekR.data")
library("MMSeekR")
data("NNscore.hg19")
data("hg19.seqLengths")
methFile <- system.file("extdata", "TCGA_BLCA_A13J_test.tab", package="MMSeekR")
if(!dir.exists("TestResult/")){dir.create("TestResult/")}
meth <- readMethylomeNew(fileName=methFile, NNdat=NNscore.hg19, seqLengths=hg19.seqLengths)
hmm.modelList=trainPMDHMDNew(meth, "chr16", 201, 1, "TestResult/TCGA_BLCA_A13J_test.multiModel.pdf")
y.list=PMDviterbiSegNew(meth, hmm.modelList, 201, 1)
}
}
