% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMethylomeNew.R
\name{readMethylomeNew}
\alias{readMethylomeNew}
\title{Load Bis-seq data}
\usage{
readMethylomeNew(fileName, NNdat, seqLengths, inputFormat = "text")
}
\arguments{
\item{fileName}{Text file or RDS file for GRanges object.}

\item{NNdat}{A dataframe of neural network score}

\item{seqLengths}{A named vector indicating the chromosome lengths of the genome used.}

\item{inputFormat}{File format. If inputFormat is set to "text" (default), the
argument FileName should refer to a tab-delimited text file in the
inputFormat: chromosome position T M, where each line stands for a CpG,
the position refers to the 1-base position of the methylated C (on the sense strand; if the methylated C was on antisense strand, please provide the position of G base), T is the
total number of reads (total counts) covering the CpG and M is the total
number of methylated counts. If inputFormat="GRanges", the file is assumed to be a
GRanges object, containing T and M as first and second data-value
entries, saved in rds format.}
}
\value{
A GRanges object containing the coordinates, total (T) and methylated counts (M)
}
\description{
Loading Bis-seq data from tab-delimited file or saved GRanges object
}
\examples{
\dontrun{
library("MMSeekR.data")
library("MMSeekR")
data("NNscore.hg19")
data("hg19.seqLengths")
methFile <- system.file("extdata", "TCGA_BLCA_A13J_test.tab", package="MMSeekR")
meth <- readMethylomeNew(fileName=methFile, NNdat=NNscore.hg19, seqLengths=hg19.seqLengths)
}
}
