% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSingleModel.R
\name{runSingleModel}
\alias{runSingleModel}
\title{Predict PMDs using single-model (the origin model of MethylSeekR)}
\usage{
runSingleModel(genomeVersion, fileName, outputIndex, trainChr="chr22", inputFormat="text", num.cores = 1, nCGbin = 101, rmSNP=FALSE, SNPFile=NULL)
}
\arguments{
\item{genomeVersion}{genome version. hg19 or hg38 for human.}

\item{fileName}{Methylome file. Text file or RDS file for GRanges object.}

\item{outputIndex}{Ouput file index, which contains outputPath and outputPrefix, like "outputPath/outputPrefix"}

\item{trainChr}{Chromosome on which HMM should be trained. Must be one of the sequence levels in Methylome file. (default chr22)}

\item{inputFormat}{File format. If format is set to "text" (default), the
argument FileName should refer to a tab-delimited text file in the
format: chromosome position T M, where each line stands for a CpG,
the position refers to the 1-base position of the methylated C (on the sense strand; if the methylated C was on antisense strand, please provide the position of G base), T is the
total number of reads (total counts) covering the CpG and M is the total
number of methylated counts. If format="GRanges", the file is assumed to be a
GRanges object, containing T and M as first and second data-value
entries, saved in rds format.}

\item{num.cores}{The number of cores used for the calculations (default 1)}

\item{nCGbin}{The number of CpGs in each sliding window used to calculate alpha (default 101). The default is highly recommended. This number should be an odd number}

\item{rmSNP}{Whether to remove SNPs (default FALSE). If yes, SNPFile should not be NULL.}

\item{SNPFile}{A text file for SNP file. This file should refer to a tab-delimited text file in the format: chromosome position, where each line stands for a SNP. It only works when rmSNP=T.}
}
\value{
No return value
}
\description{
PMD prediction using single-model
}
\examples{
\dontrun{
library("MMSeekR.data")
library("MMSeekR")
methFile <- system.file("extdata", "TCGA_BLCA_A13J_test.tab", package="MMSeekR.data")
## Do not remove SNPs
runSingleModel("hg19", methFile, "TestResults/TCGA_BLCA_A13J_test")
## Remove SNPs
SNPFile <- system.file("extdata", "common_chr16_22_20180418.tab", package="MMSeekR")
runSingleModel("hg19", methFile, "TestResults/TCGA_BLCA_A13J_test", rmSNP=TRUE, SNPFile=SNPFile)
}
}
