% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePMDSegRmBlackListNew.R
\name{savePMDSegRmBlackListNew}
\alias{savePMDSegRmBlackListNew}
\title{Remove blackList and save PMD segments}
\usage{
savePMDSegRmBlackListNew(seg, blackList, GRangesFilename = NULL, TableFilename = NULL)
}
\arguments{
\item{seg}{GRanges object.}

\item{blackList}{A tibble of blackList}

\item{GRangesFilename}{Filename of the GRanges object.}

\item{TableFilename}{Filename of the PMD table.}
}
\value{
No return value
}
\description{
Remove blackList and save PMD segments in rds format and a tab-delimited file.
}
\examples{
\dontrun{
library("MMSeekR.data")
library("MMSeekR")
data("NNscore.hg19")
data("hg19.seqLengths")
data("hg19.blackList")
methFile <- system.file("extdata", "TCGA_BLCA_A13J_test.tab", package="MMSeekR")
meth <- readMethylomeNew(fileName=methFile, NNdat=NNscore.hg19, seqLengths=hg19.seqLengths)
if(!dir.exists("TestResult/")){dir.create("TestResult/")}
hmm.modelList=trainPMDHMDNew(meth, "chr16", 201, 1, "TestResult/TCGA_BLCA_A13J_test.multiModel.pdf")
y.list=PMDviterbiSegNew(meth, hmm.modelList, 201, 1)
segments = createGRangesObjectPMDSegNew(meth, y.list, 1, hg19.seqLengths)
savePMDSegRmBlackListNew(segments, hg19.blackList, "TestResult/TCGA_BLCA_A13J_test.multiModel.rds", "TestResult/TCGA_BLCA_A13J_test.multiModel.PMDs.bed")
}
}
