% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainPMDHMDNew.R
\name{trainPMDHMDNew}
\alias{trainPMDHMDNew}
\title{Training PMD and HMD models}
\usage{
trainPMDHMDNew(m, chr.sel, nCGbin=201, num.cores=1, pdfFilename)
}
\arguments{
\item{m}{A GRanges object containing the coordinates, total (T) and methylated counts (M)}

\item{chr.sel}{Chromosome on which HMM should be trained. Must be one of the sequence levels of m. (default chr16)}

\item{nCGbin}{The number of CpGs in each sliding window used to calculate PCC correlation and Mvalue (default 201). The default is highly recommended.}

\item{num.cores}{The number of cores used for the calculations (default 1)}

\item{pdfFilename}{Name of the pdf file in which the figure is saved.}
}
\value{
A list of the parameters of the HMM model
}
\description{
Train a two-state Hidden Markov Model (HMM) with Gaussian emissions.
}
\examples{
\dontrun{
library("MMSeekR.data")
library("MMSeekR")
data("NNscore.hg19")
data("hg19.seqLengths")
methFile <- system.file("extdata", "TCGA_BLCA_A13J_test.tab", package="MMSeekR")
if(!dir.exists("TestResult/")){dir.create("TestResult/")}
meth <- readMethylomeNew(fileName=methFile, NNdat=NNscore.hg19, seqLengths=hg19.seqLengths)
hmm.modelList=trainPMDHMDNew(meth, "chr16", 201, 1, "TestResult/TCGA_BLCA_A13J_test.multiModel.pdf")
}
}
